Change locations actions/reducer to restaurants
